+++ backend/constants.ts
@@ -1,0 +1,8 @@
+
+const LY_INCLUDE_FOLDERS = ["../ly"];
+
+
+
+export {
+	LY_INCLUDE_FOLDERS,
+};