+++ inc/lilyParser/lilyDocument.ts
@@ -7,7 +7,7 @@ import {
 	parseRaw,
 	BaseTerm, Assignment, LiteralString, Command, Variable, MarkupCommand, Grace, AfterGrace, Include, Version, Block, InlineBlock,
 	Scheme, Chord, BriefChord, MusicBlock, SimultaneousList, ContextedMusic, Divide, Tempo, PostEvent, Primitive, ChordElement, MusicEvent,
-	SchemePointer, Comment, Language, StemDirection,
+	SchemePointer, Comment, Language, StemDirection, Repeat,
 } from "./lilyTerms";
 import LilyInterpreter from "./lilyInterpreter";
 
@@ -836,5 +836,15 @@ export default class LilyDocument {
 
 			block.body = block.body.filter(term => !redundants.includes(term));
 		});
+	}
+
+
+	removeRepeats () {
+		this.root.forEachTerm(MusicBlock, block => block.body = block.body.map(term => {
+			if (term instanceof Repeat)
+				return term.bodyBlock;
+
+			return term;
+		}));
 	}
 };