+++ changelog.md
@@ -2,6 +2,7 @@
 ## TODO
 	implicit notes matching with sheet document
 		match lilyNotation with articulated MIDI (ignored repeats), then unfold repeats on lilyNotation, include D.C., D.S.
+		parsing repeat mark among measure index
 	staffNotation
 		remove note time from staff document
 	lilyInterpreter
@@ -22,6 +23,8 @@
 
 
 *	2020.9.4
+
+	+	lilyNotation: class of Notation added.
 
 	+	lilyNotation: added measureHeads on Notation.